refactor:#39 폴더명 수정
