+++ src/components/features/mypage/myReserveDetail/reserveContents/InfoDetailContainer.tsx
