+++ src/components/features/mypage/myReserveDetail/reserveContents/InfoMap.tsx
