+++ src/components/features/mypage/myReserveDetail/reserveContents/InfoTitleBox.tsx
